requirements.txt
# Simple dependencies for Qwen agent
requests>=2.25.0
python-dotenv>=0.19.0
